+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning].
 - ([#2046]) Drop support for ESLint v8 and v9.
 - ([#2046]) Drop support for Node.js 18, 20.0.0 through 20.18.x, and 22.0.0
   through 22.12.x.
+- ([#2047]) Drop various allowed values for the `allowed` option of the `no-top-level-variables` rule.
 
 ## 3.6.3 (2026-06-25)
 
@@ -276,3 +277,4 @@ Versioning].
 [#1884]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1884
 [#1947]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1947
 [#2046]: https://github.com/ericcornelissen/eslint-plugin-top/pull/2046
+[#2047]: https://github.com/ericcornelissen/eslint-plugin-top/pull/2047